* Fixes for the 'Requires' tooltip showing weird Squirrel function names (caused by the usual try/catch decompile failure) * Completed 'Persistent Abilities'. The opens the way for long running status effects such as XP tomes. Also means buffs won't be lost across relogs in dungeons. Downside (for some players anyway), logging out will no longer clear debuffs :) * Doing the work for persistent abilities (and so long running effects), highlighted a few problems with casting of times losing precision. Can now easily have 30 day buffs. * Fixed some problems with CTF script (leaving party/quitting game when leaving instance and dropping the right flag).

+++ SOURCE/Server/AbilityTime.h
@@ -11,7 +11,7 @@ struct ActiveBuff
 	unsigned char buffType;
 	short abID;
 	short abgID;
-	int durationS;
+	double durationS;
 	unsigned long castEndTimeMS;
 };
